Refactoring the implementation of type conversion
