+++ core/src/main/scala/org/apache/spark/sql/execution/datasources/hbase/HBaseFilter.scala
@@ -135,7 +135,7 @@ object HBaseFilter extends Logging{
   }
 
   private def toBytes[T](value: T, att: String, relation: HBaseRelation): Array[Byte] = {
-    Utils.toBytes(value, relation.getField(att))
+    TypeManager(relation.getField(att)).toBytes(value)
   }
